real-time-review-adding: move event into separate folder
