+++ app/Actions/Review/CreateReviewAction.php
@@ -8,7 +8,7 @@ use Hedonist\Repositories\User\UserRepositoryInterface;
 use Hedonist\Repositories\Place\PlaceRepositoryInterface;
 use Hedonist\Repositories\Review\ReviewRepositoryInterface;
 use Hedonist\Exceptions\Place\PlaceDoesNotExistException;
-use Hedonist\Events\ReviewAddEvent;
+use Hedonist\Events\Review\ReviewAddEvent;
 
 class CreateReviewAction
 {
